+++ src/assets.rs
@@ -1,6 +1,9 @@
 use log::info;
 use rust_embed::RustEmbed;
-use std::collections::HashSet;
+use std::collections::{HashMap, HashSet};
+use std::fs::File;
+use std::io::{BufRead, BufReader, Write};
+use std::path::Path;
 
 #[derive(RustEmbed)]
 #[folder = "data/"]
@@ -8,18 +11,22 @@ struct Asset;
 
 #[derive(Debug)]
 pub struct Dict {
+    pub path: String,
     pub common_words: HashSet<String>,
     pub species_black: HashSet<String>,
     pub uninomials_black: HashSet<String>,
+    pub genera: HashSet<String>,
 }
 
 impl Dict {
-    pub fn new() -> Self {
+    pub fn new(path: &str) -> Self {
         info!("Building temp dictionaries");
         let mut dict = Dict {
+            path: path.to_owned(),
             common_words: HashSet::new(),
             species_black: HashSet::new(),
             uninomials_black: HashSet::new(),
+            genera: HashSet::new(),
         };
 
         let common_words =
@@ -46,6 +53,67 @@ impl Dict {
             dict.species_black.insert(word.trim().to_owned());
         }
 
+        let f = File::open(Path::new(path).join("genera.txt")).unwrap();
+        let r = BufReader::new(f);
+        for line in r.lines() {
+            let line = line.unwrap().to_owned();
+            dict.genera.insert(line.trim().to_owned());
+        }
+
         dict
+    }
+
+    pub fn canonicals(&self) {
+        info!("Creating interim csv files.");
+        let mut uninomials: HashMap<String, u32> = HashMap::new();
+        let mut genera: HashMap<String, u32> = HashMap::new();
+        let mut species: HashMap<String, u32> = HashMap::new();
+        let f = File::open(Path::new(&self.path).join("names.txt")).unwrap();
+        let r = BufReader::new(f);
+        for line in r.lines() {
+            let line: String = line.unwrap().to_owned();
+            if line.find("×").is_some() {
+                continue;
+            }
+            let words: Vec<&str> = line.trim().split(" ").collect();
+            if words.len() == 1 {
+                let uni = words[0];
+                if self.genera.get(uni).is_none() {
+                    let entry = uninomials.entry(words[0].to_owned()).or_default();
+                    *entry += 1;
+                } else {
+                    let entry = genera.entry(words[0].to_owned()).or_default();
+                    *entry += 1;
+                }
+            } else {
+                let entry = genera.entry(words[0].to_owned()).or_default();
+                *entry += 1;
+                for word in &words[1..] {
+                    let word = word.to_owned();
+                    let entry = species.entry(word.to_owned()).or_default();
+                    *entry += 1;
+                }
+            }
+        }
+        let mut genera_keys: Vec<String> = Vec::new();
+        for (k, v) in &uninomials {
+            if genera.contains_key(k) {
+                let entry = genera.entry(k.clone()).or_default();
+                *entry += *v;
+                genera_keys.push(k.to_owned());
+            }
+        }
+        for k in genera_keys {
+            uninomials.remove(&k);
+        }
+        self.build_csv("uninomials.csv", uninomials);
+        self.build_csv("genera.csv", genera);
+        self.build_csv("species.csv", species);
+    }
+    fn build_csv(&self, name: &str, data: HashMap<String, u32>) {
+        let mut f = File::create(Path::new(&self.path).join(name)).unwrap();
+        for (k, v) in data {
+            writeln!(f, "{},{}", k, v).unwrap();
+        }
     }
 }